# domain-banking-cards-sdk/src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:18080
  description: Generated server url
...
paths:
...
  /api/v1/cards:
    post:
      tags:
      - Cards
      summary: Issue a new card
      description: Issues a new debit or credit card for a customer
      operationId: issueCard
      parameters:
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCardCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IssueCardResponse"
  /api/v1/cards/{cardId}/virtual:
    post:
      tags:
      - Virtual Cards
      summary: Issue virtual card
      description: Issues a virtual card linked to a physical card
      operationId: issueVirtualCard
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueVirtualCardCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/VirtualCardResponse"
  /api/v1/cards/{cardId}/unblock:
    post:
      tags:
      - Cards
      summary: Unblock a card
      description: Removes a temporary block from a card
      operationId: unblockCard
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: unblockedBy
        in: query
        required: false
        schema:
          type: string
          default: CUSTOMER
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
...
  /api/v1/cards/{cardId}/replace:
    post:
      tags:
      - Cards
      summary: Replace a card
      description: "Replaces a lost, stolen, or damaged card"
      operationId: replaceCard
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: reason
        in: query
        required: true
        schema:
          type: string
      - name: transferLimits
        in: query
        required: false
        schema:
          type: boolean
          default: true
      - name: transferSecuritySettings
        in: query
        required: false
        schema:
          type: boolean
          default: true
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReplaceCardResponse"
...
  /api/v1/cards/{cardId}/credit-line:
    post:
      tags:
      - Credit Line
      summary: Setup credit line
      description: Sets up a revolving credit line for a credit card
      operationId: setupCreditLine
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupCreditLineCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreditLineResponse"
  /api/v1/cards/{cardId}/cancel:
    post:
      tags:
      - Cards
      summary: Cancel a card
      description: Permanently cancels a card
      operationId: cancelCard
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: reason
        in: query
        required: true
        schema:
          type: string
      - name: cancelledBy
        in: query
        required: false
        schema:
          type: string
          default: CUSTOMER
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/v1/cards/{cardId}/block:
    post:
      tags:
      - Cards
      summary: Block a card
      description: Temporarily blocks a card
      operationId: blockCard
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: reason
        in: query
        required: true
        schema:
          type: string
      - name: blockedBy
        in: query
        required: false
        schema:
          type: string
          default: CUSTOMER
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
...
  /api/v1/cards/{cardId}/activate:
    post:
      tags:
      - Cards
      summary: Activate a card
      description: Activates a card using the activation code
      operationId: activateCard
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: activationCode
        in: query
        required: true
        schema:
          type: string
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/v1/backoffice/cards/{cardId}/virtual:
    get:
      tags:
      - Card Backoffice
      summary: Get virtual cards
      description: Retrieves all virtual cards linked to the parent card
      operationId: getVirtualCards
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VirtualCardDTO"
    post:
      tags:
      - Card Backoffice
      summary: Create virtual card
      description: Creates a virtual card (saga operation)
      operationId: createVirtualCard
      parameters:
      - name: cardId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVirtualCardRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: object
...
components:
  schemas:
...
    IssueCardCommand:
      type: object
      properties:
        customerId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        cardProgramId:
          type: string
          format: uuid
        cardType:
          type: string
        currency:
          type: string
        cardNetworkId:
          type: string
          format: uuid
        issuerId:
          type: string
          format: uuid
        binId:
          type: string
          format: uuid
        cardHolderName:
          type: string
        cardHolderId:
          type: string
        isPhysical:
          type: boolean
        isVirtual:
          type: boolean
        isPrimary:
          type: boolean
    IssueCardResponse:
      type: object
      properties:
        cardId:
          type: string
          format: uuid
        executionId:
          type: string
        status:
          type: string
    IssueVirtualCardCommand:
      type: object
      properties:
        parentCardId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        purpose:
          type: string
        spendingLimit:
          type: number
        expiresAt:
          type: string
          format: date-time
        singleUse:
          type: boolean
    VirtualCardResponse:
      type: object
      properties:
        virtualCardId:
          type: string
          format: uuid
        parentCardId:
          type: string
          format: uuid
        executionId:
          type: string
        status:
          type: string
...
    ReplaceCardResponse:
      type: object
      properties:
        oldCardId:
          type: string
          format: uuid
        newCardId:
          type: string
          format: uuid
        executionId:
          type: string
        status:
          type: string
...
    SetupCreditLineCommand:
      type: object
      properties:
        cardId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        creditLimit:
          type: number
        interestRate:
          type: number
        currency:
          type: string
    CreditLineResponse:
      type: object
      properties:
        revolvingLineId:
          type: string
          format: uuid
        cardId:
          type: string
          format: uuid
        creditLimit:
          type: number
        executionId:
          type: string
        status:
          type: string
...
    CreateVirtualCardRequest:
      type: object
      properties:
        customerId:
          type: string
          format: uuid
        spendingLimit:
          type: number
...
    VirtualCardDTO:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
        dateUpdated:
          type: string
          format: date-time
        virtualCardId:
          type: string
          format: uuid
        cardId:
          type: string
          format: uuid
        deviceId:
          type: string
        virtualCardNumber:
          type: string
        virtualCardStatus:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
          - CANCELLED
        creationTimestamp:
          type: string
          format: date-time
        updateTimestamp:
          type: string
          format: date-time
...